requirements.txt
# Streamlit for UI
streamlit>=1.31.0

# LiteLLM for unified LLM interface
litellm>=1.30.0

# Environment variables
python-dotenv>=1.0.0

# Data manipulation and visualization
pandas>=2.2.0
plotly>=5.18.0

# Data validation
pydantic>=2.0.0

# Testing (optional but recommended)
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0

# Additional utilities
requests>=2.31.0
